+++ FreeRTOS-Plus/Demo/FreeRTOS_Plus_Reliance_Edge_and_CLI_Windows_Simulator/ConfigurationFiles/redconf.c
@@ -11,5 +11,5 @@
 
 const VOLCONF gaRedVolConf[REDCONF_VOLUME_COUNT] =
 {
-    { 512U, 65536U, false, 256U, "" }
+    { 512U, 65536U, false, 256U, 0, "" }
 };